Tracing C++ functions to resolving error from "::"
There are lots of C++ based 3rd party application and  inbox apps as well. So, when engineer debugs the application they might confuse how they can escape the :: probe issue as C++ use scope operator :: so hard to specify the interested function to probe. And, one of our partners brought a question to us and asked us to add this information to the online doc for them , their customer and others.  I got this information from the Microsoft internal wiki and polished little. 

https://www.osgwiki.com/wiki/DTrace_on_Windows#Tracing_functions_written_in_C.2B.2B

Feel free to polish the contents but hope to added this information here.